rename to "utils.mjs", make "build.mjs" executable
